fixed databinding

+++ app/src/main/java/by/kesso/pixabaytest/domain/entity/PixaImage.kt
@@ -5,13 +5,13 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class PixaImage(
-    val id: Int,
+    val id: Long,
 
     val type: String?,
 
     val largeImageURL: String?,
     val imageHeight: Int?,
-    val imageSize: Int?,
+    val imageSize: Long?,
     val imageWidth: Int?,
 
     val webFormatURL: String?,
@@ -24,14 +24,14 @@ data class PixaImage(
 
     val pageURL: String?,
 
-    val likes: Int?,
-    val views: Int?,
-    val downloads: Int?,
-    val comments: Int?,
+    val likes: Long?,
+    val views: Long?,
+    val downloads: Long?,
+    val comments: Long?,
     val tags: String?,
-    val collections: Int?,
+    val collections: Long?,
 
     val user: String?,
-    val userId: Int?,
+    val userId: Long?,
     val userImageURL: String?,
 ): Parcelable
